cleanup in preparation of batch reader implementation
